server/requirements.txt
fastapi
uvicorn
langchain
sentence-transformers
chromadb
python-dotenv
PyPDF2
